LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/specifier-error-expected.txt

CONSOLE MESSAGE: TypeError: Module name, 'string-without-dot-slash-prefix' does not resolve to a valid URL.

PASS Test that invalid module specifier leads to TypeError on window.
